Added some classes"
the VMWriter Skeleton that we call inside the CompilerEngine when we
want to generate the VM code and write to the file.

The SymbolTable, last 3 methods need to be completed. This data
structure makes use of the custom built SymbolRecord class that holds
all three fields for each symbol identifier.

KIND was made into an Enum for easier access and so that some of the
methods suggested in ECS could actually return something useful rather
than a variable type.